add /synced so a caller can wait for its own commit to be searchable
The one guarantee an in-database extension gives that an external pipeline does
not: the commit returned and the row is already in search. ZomboDB gets it by
living inside the transaction and pays by putting the search engine in every
write path. This gets it by letting the read that needs it wait, so a write
never pays and an outage in the target never blocks a commit.

GET /synced waits until the acknowledged position passes the source's, and with
refresh=true until the target has made the writes searchable — stored is not
searchable on OpenSearch until a refresh, which is the gap the demo app had
been papering over with a polling loop. The wait is driven by the watch channel
the sink task already updates, so a caller is woken by the write itself.

Two source realities had to be handled for this to work rather than merely
appear to. PostgreSQL skips transactions touching no published table, so the
position a caller reads includes activity the pipeline will never see and the
gap never closes on a quiet database: measured stuck for 30s while the row was
already indexed. The endpoint now emits a logical decoding message when it
finds itself behind — a marker the stream does carry, no table, no DDL. MySQL's
binlog is server-wide so the commit's own XID reaches the position; a heartbeat
period covers a genuinely idle server.

position is optional and reading it is better left to us: it requires
REPLICATION CLIENT on MySQL, which an application account should not hold.

Also fixes two things found on the way. The commit position acknowledged is now
end_lsn rather than the commit record's own offset, which is what
pg_current_wal_lsn() reports and what lets the server release that record. And
child lookups compare integers as ::bigint: binding an i64 against an int4
column failed outright, a regression from removing the text cast that the
suite's bigint-only schema had missed. The dev schema now has a narrower child
key so it cannot happen again.

Verified 20 writes on PostgreSQL and 15 on MySQL, each followed immediately by
a search with no sleep and no retry: zero misses, zero timeouts, /synced
returning in 5 ms at the median and 25 ms at worst. Both suites assert the
strict form. PostgreSQL 40/40, MySQL 8.0 and MariaDB 11.8 25/25 each.

+++ dev/seed.sql
@@ -26,9 +26,11 @@ CREATE TABLE IF NOT EXISTS orders (
     total       numeric(12, 2) NOT NULL
 );
 
+-- integer rather than bigint on purpose: a child key narrower than i64 is the
+-- common case and once broke the typed comparison
 CREATE TABLE IF NOT EXISTS tickets (
     id          bigint PRIMARY KEY,
-    customer_id bigint NOT NULL REFERENCES customers(id),
+    customer_id integer NOT NULL REFERENCES customers(id),
     subject     text NOT NULL
 );
